[feat/#46]리뷰 crud 완료

+++ demo/src/main/java/com/example/moviezip/dao/ReviewDao.java
@@ -11,11 +11,11 @@ public interface ReviewDao {
 
     List<Review> getMyReviewList(String userId) throws DataAccessException;
 
-    void deleteReview(long rvId) throws DataAccessException;
+    void deleteReview(int rvId) throws DataAccessException;
 
     List<Review> getReviewList(long mvId) throws DataAccessException;
 
-    Review getReviewDetail(long rvId) throws DataAccessException;
+    Review getReviewDetail(int rvId) throws DataAccessException;
 
     int updateReview(Review review) throws DataAccessException;
